adding routes, nav comp
