+++ seans_clothing/src/routes/sign-in/sign-in.component.jsx
@@ -1,0 +1,9 @@
+const SignIn = () => {
+    return (
+        <div>
+            <h1>Sign In Page</h1>
+        </div>
+    )
+}
+
+export default SignIn;
